support read-only mode

+++ pkg/meta/interface.go
@@ -222,7 +222,7 @@ type Meta interface {
 	// Create creates a file in a directory with given name.
 	Create(ctx Context, parent Ino, name string, mode uint16, cumask uint16, inode *Ino, attr *Attr) syscall.Errno
 	// Open checks permission on a node and track it as open.
-	Open(ctx Context, inode Ino, flags uint8, attr *Attr) syscall.Errno
+	Open(ctx Context, inode Ino, flags uint32, attr *Attr) syscall.Errno
 	// Close a file.
 	Close(ctx Context, inode Ino) syscall.Errno
 	// Read returns the list of slices on the given chunk.
